Enhanced many plugins and introduced some new features
mattata v4.4! More administration!

You can now use /warn in reply to a user in your group to warn them
and, if they reach 3 warnings, they will automatically be banned!

Added a new 'report' plugin. When a user executes /report or /ops in
your group it will alert all admins in the group with an optional
message that the user can give as an argument!

Small improvements to the calc plugin - the '÷' symbol is now
recognised as a division symbol!

Added a copypasta plugin which riddles the replied-to message with
cancerous emoji, thanks to @JuanPotato for helping me with the python

Added an inline version of the last.fm plugin, respective help
documentation and added /start as an alias of /help

The rules plugin now uses 3 separate commands rather than arguments,
for ease of use

Added a link plugin which allows a group admin to set the group link
for users to easily retrieve using /link - see the documentation of
this command (using /help link) for more information

The translation plugin can now translate any language into any other
language, using /translate <language> <text> instead of /translate
<text>

... and several other backend changes which users probably won't notice!