[v9] The "administration" update (amongst other things)
Well, it is safe to say that this is one of the biggest updates so far!

Here are some of the changes:

- Added /bugreport, /bigtext, /fortune, /google, /netflix, /slap
- Improved most plugins to a certain extent
- Re-wrote the entire administration functionality, now including an
interactive menu accessible via /administration
- giphy -> gif
- Re-wrote copypasta.py in lua and included it in the relevant plugin,
same with lyrics.py - meaning Python is no longer a dependency for
mattata
- Added support to /author for all file types
- Moderation commands; give a user administrative powers with /mod, and
strip them of it with /demod
- Set a pinned message that can easily be edited with /pin
- All data is stored in redis now, deprecated users.json and groups.json
- Create custom #hashtag commands with /custom, which sends the given
text every time somebody uses the said hashtag
- Improved anti-spam, and added anti-bot functionality
- Added an admins-only mode (toggle-able from /administration), which
(if enabled) only allows admins to use mattata in the chat
- /report now forwards the replied-to message to all admins via private
chat, and provides a direct link to the reported message
- A button to view the rules in private chat is accessible underneath
the welcome message upon somebody joining the chat
- Began changing the way the language files are implemented, but this
isn't my priority right now
- Made configuration.lua neater and removed clutter
- Administration is disabled by default and can be enabled with
/administration
- Fixed an insane amount of bugs, a lot of which could have probably
been avoided in the first place, if I didn't rush to meet deadlines!

+ tons more!