[v19] Installation improvements, more regular garbage collection and plugin enhancements
Check the official channel (https://t.me/mattata) for the latest news
about features and updates, as and when they happen!

It is highly recommended you run the appropriate installation script
after updating!

I’ve also commented on a lot more of the framework, and I’ll be doing
that more often to make things easier to understand!

A big shoutout to t.me/LKD70 for kindly donating (not once, but twice);
and for adding the correct links to get a key for each API in the
configuration file!